Refactor doctors list view & implement the api data to the list
